Update call state handling mechanism to access finished call state
This commit updates call state handling mechanism. Instead of removing
call state from call stack and serializing it, the call state is now
also tracked by a map which is emptied only at the tracing end.
This ensures that the call state can be accessed even after the call
has finished. This is the first step towards escaped promise detection.